requirements.txt

streamlit
geopandas
shapely
python-docx
pandas
fiona
pyproj
rtree
google-generativeai
